MEC for NCHW
Works faster if image is bigger than the number of patches.
Otherwise, the overhead of extra Kernel reshaping (as compared to
im2col) slows down.
Does not work for M > 1.

Found that convGemm computes incorrect result for some sizes (the same
fix for its GEMM structure as in the last few commits)